move ascendingDefined
